Extract 'sample' fixture
